chore: Complete OpenSpec validation and code quality improvements
Complete Phase 6 (Testing & Validation) and Phase 7 (Documentation) of the
fix-gl-renderer-bcf-serialization OpenSpec change.

Code Quality Fixes:
- Fix clippy warnings: remove unused enumerate(), unnecessary casts
- Add `#[allow(dead_code)]` to bcf_cpu_tracer (WIP module)
- Add safety documentation to all unsafe functions in gl_tracer and gpu_tracer
- Fix empty lines after doc comments

Documentation Enhancements:
- Expand GL tracer module docs with detailed BCF serialization approach
- Document GPU upload strategy (1D-like 2D texture, R8UI format)
- Add inline comments explaining BCF node format and coordinate systems
- Update OpenSpec tasks.md to mark all phases complete

Testing Validation:
- Verified renderer runs successfully with BCF serialization (21 bytes)
- Confirmed shader compilation and GPU buffer upload working
- Validated clippy passes with -D warnings
- Confirmed cargo fmt compliance

All phases (1-7) of the BCF GL renderer fix are now complete.

+++ crates/cube/src/mesh/generator.rs
@@ -168,13 +168,7 @@ pub fn generate_face_mesh<B, F>(
                 let uvs = face_info
                     .face
                     .uvs(position.x, position.y, position.z, size, uv_scale);
-                builder.add_textured_face(
-                    vertices,
-                    normal,
-                    color,
-                    uvs,
-                    face_info.material_id as u8,
-                );
+                builder.add_textured_face(vertices, normal, color, uvs, face_info.material_id);
             } else {
                 // Solid color material (0-1, 128-255)
                 builder.add_face(vertices, normal, color);
